user-portal/pyproject.toml
[tool.poetry]
name = "arxiv-user-portal"
version = "0.1.0"
description = "arXiv user admin API"
authors = ["Naoyuki Tai <ntai@arxiv.org>"]
license = "BSD-3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pydantic = {extras = ["email"], version = "^2.10.5"}
arxiv-base = {git = "https://github.com/arXiv/arxiv-base.git", rev = "1eef49b34b30bb0fc2b8e137885d3babf570b989"}
pymysql = "^1.1.0"
ruamel-yaml = "^0.18.6"
python-json-logger = "^2.0.7"
python-dateutil = "^2.8.2"
pyjwt = "^2.9.0"
requests = "^2.32.3"
jwcrypto = "^1.5.6"
httpx = "^0.27.2"
cachetools = "^5.5.0"
sqlalchemy = "^2.0.36"
python-multipart = "^0.0.20"
flask-wtf = "^1.2.2"
bootstrap-flask = "^2.4.1"
flask-sqlalchemy = "^3.1.1"
pycountry = "^24.6.1"
captcha = "^0.6.0"
fastapi = "^0.115.8"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.34.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
